requirements.txt
alpha-vantage==2.1.3
aniso8601==8.0.0
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-RESTful==0.3.8
gcloud==0.17.0
googleapis-common-protos==1.51.0
httplib2==0.17.0
itsdangerous==1.1.0
Jinja2==2.11.1
jws==0.1.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
numpy==1.18.1
oauth2client==3.0.0
pandas==1.0.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
PyJWT==1.4.2
pyparsing==2.4.6
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2019.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
Werkzeug==1.0.0
